stroke_plugin.c: fix segfault under musl, NDMS-1113
(gdb) bt
 #0  0x77fa1b6c in strlen () from /lib/ld-musl-mips-sf.so.1

 #1  0x77eb15cc in builtin_vsnprintf (buffer=0x778869a8 "0\n123\n(null)\n",
     n=8192, format=0x77e0b330 "%d\n%d\n%s\n%s\n%u\n", ap=0x778889f0)
     at utils/printf_hook/printf_hook_builtin.c:847
 #2  0x77eb21b0 in builtin_vfprintf (stream=0x779b2ad0,
     format=0x77e0b330 "%d\n%d\n%s\n%s\n%u\n", ap=0x778889e0)
     at utils/printf_hook/printf_hook_builtin.c:1235

+++ src/libcharon/plugins/stroke/stroke_plugin.c
@@ -156,7 +156,9 @@ static void update_connections(void)
 			ike_version = peer_cfg->get_ike_version(peer_cfg);
 			my_addr = ike_cfg->get_my_addr(ike_cfg);
 			other_addr = ike_cfg->get_other_addr(ike_cfg);
-			fprintf(fd, "%d\n%d\n%s\n%s\n%u\n", time(NULL), time_monotonic(NULL),
+			fprintf(fd, "%lld\n%lld\n%s\n%s\n%u\n",
+				(long long)time(NULL),
+				(long long)time_monotonic(NULL),
 				my_addr, other_addr, ike_version);
 
 			if (ike_version == IKEV1 && peer_cfg->use_aggressive(peer_cfg))
@@ -422,7 +424,9 @@ static void update_sa(void)
 
 		if (fd != NULL)
 		{
-			fprintf(fd, "%d\n%d\n", time(NULL), time_monotonic(NULL));
+			fprintf(fd, "%lld\n%lld\n",
+				(long long)time(NULL),
+				(long long)time_monotonic(NULL));
 
 			fprintf(fd, "ikesa\n");
 			update_ike_sa(fd, ike_sa);
